Added search functionality per subject
