+++ src/Science.js
@@ -3,7 +3,6 @@ import MainHeader from './Components/MainHeader';
 import CategoryHeader from './Components/CategoryHeader';
 import CategoryBannerScience from './Components/CategoryBannerScience';
 import Search from './Components/Search';
-import { Link } from 'react-router-dom';
 import './Styles/Page.css';
 
 const Science = () => {
@@ -14,7 +13,7 @@ const Science = () => {
           <MainHeader></MainHeader>
           <CategoryHeader></CategoryHeader>
           <CategoryBannerScience></CategoryBannerScience>
-          <Search></Search>
+          <Search subject='Science'></Search>
         </div>
       </body>
     </div>
